Moved examples to top level
