+++ .storybook/config.js
@@ -1,6 +1,6 @@
 import { configure } from '@storybook/html';
 
-const stories = require.context('../src/examples', true, /\.stories\.js$/);
+const stories = require.context('../examples', true, /\.stories\.js$/);
 
 function loadStories() {
     stories.keys().forEach(fileName => stories(fileName));
