+++ examples/javascript.stories.js
@@ -1,33 +1,36 @@
-
 import { storiesOf } from '@storybook/html';
-import copyCodeBlock from '../copyCodeBlock';
-import { customStyles, customHtml, opts } from './customHtml';
-import { usageExample, usageExampleJsHighlight } from '../helpers';
+import copyCodeBlock from '../src/copyCodeBlock';
+import { customStyles } from './customHtml';
+import { usageExample, usageExampleJsHighlight } from './helpers';
 import hljs from 'highlight.js/lib/highlight';
 
 // Register languages for hljs
-hljs.registerLanguage('html', require('highlight.js/lib/languages/xml'));
 hljs.registerLanguage('js', require('highlight.js/lib/languages/javascript'));
 
 const a11yLightStyle = require('!url-loader!highlight.js/styles/a11y-light.css');
 const draculaStyle = require('!url-loader!highlight.js/styles/dracula.css');
 
-const htmlExample = `
-<div class='cool-container'>
-    <p class='cool-description'>Get ready for...</p>
-    <ul class='cool-list'>
-        <li class='cool-item'>rainbows</li>
-        <li class='cool-item'>and</li>
-        <li class='cool-item'>unicorns</li>
-    </ul>
-</div>`;
+const jsExample = `
+import { Pants } from 'fancy-things';
 
-storiesOf('HTML', module)
+var cool = 'cool';
+let stuff = 'stuff';
+const things = 'things';
+
+function coolStuffAndThings(a, b, c) {
+    return \`\$\{a\}\ \$\{b\}\ and \$\{c\}\`;
+}
+
+coolStuffAndThings(cool, stuff, things);
+Pants();
+`;
+
+storiesOf('Javascript', module)
     .add('Default', () => `
         <link rel="stylesheet" href="${draculaStyle}">
         <h1>Default</h1>
         <h2>Example Code</h2>
-        ${copyCodeBlock(htmlExample) /* Will not run through hljs */}
+        ${copyCodeBlock(jsExample) /* Will not run through hljs */}
         <h2>Usage</h2>
         ${copyCodeBlock(usageExample(), usageExampleJsHighlight)}
     `)
@@ -35,7 +38,7 @@ storiesOf('HTML', module)
         <link rel="stylesheet" href="${draculaStyle}">
         <h1>Custom styles, no syntax highlighting</h1>
         <h2>Example Code</h2>
-        ${copyCodeBlock(htmlExample, customStyles) /* Still not run through hljs */}
+        ${copyCodeBlock(jsExample, customStyles) /* Still not run through hljs */}
         <h2>Usage</h2>
         ${copyCodeBlock(usageExample(customStyles), usageExampleJsHighlight)}
     `)
@@ -51,15 +54,15 @@ storiesOf('HTML', module)
             from highlight.js, by adding it to the HTML's stylesheets.
         </p>
         <h2>Example Code</h2>
-        ${copyCodeBlock(htmlExample, {lang: 'html'}) /* Run through hljs, no custom styles */}
+        ${copyCodeBlock(jsExample, {lang: 'js'}) /* Run through hljs, no custom styles */}
         <h2>Usage</h2>
-        ${copyCodeBlock(usageExample({lang: 'html'}), usageExampleJsHighlight)}
+        ${copyCodeBlock(usageExample({lang: 'js'}), usageExampleJsHighlight)}
     `)
     .add('Syntax highlighting & custom styles', () => `
         <link rel="stylesheet" href="${a11yLightStyle}">
         <h1>Syntax highlighting & custom styles</h1>
         <h2>Example Code</h2>
-        ${customHtml(htmlExample) /* Run through hljs, w/custom styles */}
+        ${copyCodeBlock(jsExample, usageExampleJsHighlight) /* Run through hljs, w/custom styles */}
         <h2>Usage</h2>
-        ${copyCodeBlock(usageExample(opts), usageExampleJsHighlight)}
+        ${copyCodeBlock(usageExample(usageExampleJsHighlight), usageExampleJsHighlight)}
     `);
